+++ tools/geneos/cmd/protect.go
@@ -27,13 +27,12 @@ var protectCmd = &cobra.Command{
 	GroupID: GROUP_MANAGE,
 	Short:   "Mark instances as protected",
 	Long: strings.ReplaceAll(`
-Mark matcing instances as protected.
+Mark matching instances as protected. Various operations that affect
+the state or availability of an instance will be prevented if it is
+marked |protected|.
 
 To reverse this you must use the same command with the |-U| flag.
-There is no |unprotect| command. This is intentional.
-
-Note that you can also manually add or remove the |protected| setting
-in an instance configuration file.
+There is no |unprotect| command. This is by design.
 `, "|", "`"),
 	SilenceUsage: true,
 	Annotations: map[string]string{