+++ tools/geneos/cmd/ps.go
@@ -59,6 +59,7 @@ func init() {
 	GeneosCmd.AddCommand(psCmd)
 
 	psCmd.Flags().BoolVarP(&psCmdShowFiles, "files", "f", false, "Show open files")
+	psCmd.Flags().MarkHidden("files")
 	psCmd.Flags().BoolVarP(&psCmdJSON, "json", "j", false, "Output JSON")
 	psCmd.Flags().BoolVarP(&psCmdIndent, "pretty", "i", false, "Output indented JSON")
 	psCmd.Flags().BoolVarP(&psCmdCSV, "csv", "c", false, "Output CSV")
@@ -69,9 +70,14 @@ func init() {
 var psCmd = &cobra.Command{
 	Use:     "ps [flags] [TYPE] [NAMES...]",
 	GroupID: GROUP_VIEW,
-	Short:   "List process information for instances, optionally in CSV or JSON format",
+	Short:   "Show running instances",
 	Long: strings.ReplaceAll(`
-Show the status of the matching instances.
+The |ps| command will report details of matching and running instances.
+
+The default output is a table format intended for humans but this can
+be changed to CSV format using the |--csv|/|-c| flag or JSON with the
+|--json|/|-j| or |--pretty|/|-i| options, the latter option
+formatting the output over multiple, indented lines.
 `, "|", "`"),
 	Aliases:      []string{"status"},
 	SilenceUsage: true,
@@ -85,6 +91,9 @@ Show the status of the matching instances.
 	},
 }
 
+// CommandPS writes running instance information to STDOUT
+//
+// XXX relies on global flags
 func CommandPS(ct *geneos.Component, args []string, params []string) (err error) {
 	switch {
 	case psCmdJSON, psCmdIndent:
@@ -136,37 +145,9 @@ func psInstancePlain(c geneos.Instance, params []string) (err error) {
 
 	fmt.Fprintf(psTabWriter, "%s\t%s\t%s\t%d\t%v\t%s\t%s\t%s\t%s:%s\t%s\n", c.Type(), c.Name(), c.Host(), pid, ports, username, groupname, mtime.Local().Format(time.RFC3339), base, underlying, c.Home())
 
-	// if psCmdShowFiles {
-	// 	// list open files (test code)
-
-	// 	instdir := c.Home()
-	// 	files := instance.Files(c)
-	// 	fds := make([]int, len(files))
-	// 	i := 0
-	// 	for f := range files {
-	// 		fds[i] = f
-	// 		i++
-	// 	}
-	// 	sort.Ints(fds)
-	// 	for _, n := range fds {
-	// 		fdPath := files[n].FD
-	// 		perms := ""
-	// 		p := files[n].FDMode & 0700
-	// 		log.Debug().Msgf("%s perms %o", fdPath, p)
-	// 		if p&0400 == 0400 {
-	// 			perms += "r"
-	// 		}
-	// 		if p&0200 == 0200 {
-	// 			perms += "w"
-	// 		}
-
-	// 		path := files[n].Path
-	// 		if strings.HasPrefix(path, instdir) {
-	// 			path = strings.Replace(path, instdir, ".", 1)
-	// 		}
-	// 		fmt.Fprintf(psTabWriter, "\t%d:%s (%d bytes) %s\n", n, perms, files[n].Stat.Size(), path)
-	// 	}
-	// }
+	if psCmdShowFiles {
+		listOpenFiles(c)
+	}
 	return
 }