+++ tools/geneos/cmd/revert.go
@@ -49,12 +49,25 @@ func init() {
 var revertCmd = &cobra.Command{
 	Use:     "revert [TYPE] [NAME...]",
 	GroupID: GROUP_CONFIG,
-	Short:   "Revert migration of .rc files from backups",
+	Short:   "Revert earlier migration of configuration files",
 	Long: strings.ReplaceAll(`
-Revert migration of legacy .rc files to JSON if the .rc.orig backup
-file still exists. Any changes to the instance configuration since
-initial migration will be lost as the contents of the .rc file is
-never changed.
+The command will revert the |.rc.orig| suffixed configuration file
+for all matching instances.
+
+For any instance that is |protected| this will fail and an error reported.
+
+The original file is never updated and any changes made since the
+original migration will be lost. The new configuration file will be
+deleted.
+
+If there is already a configuration file with a |.rc| suffix then the
+command will remove any |.rc.orig| and new configuration files while
+leaving the existing file unchanged.
+
+If called with the |--executables|/|-X| option then instead of
+instance configurations the command will remove any symbolic links
+from legacy |ctl| command in |${GENEOS_HOME}/bin| that point to the
+command.
 `, "|", "`"),
 	SilenceUsage: true,
 	Annotations: map[string]string{
@@ -71,6 +84,10 @@ never changed.
 }
 
 func revertInstance(c geneos.Instance, params []string) (err error) {
+	if instance.IsProtected(c) {
+		return geneos.ErrProtected
+	}
+
 	// if *.rc file exists, remove rc.orig+new, continue
 	if _, err := c.Host().Stat(instance.ComponentFilepath(c, "rc")); err == nil {
 		// ignore errors